Initial commit for Operator Framework refactor
This commit is a refactor of Foghorn's APIs and controllers. Prior to
this, they were implemented using kubebuilder, which is overly complex
for our purposes here.

Refactoring Foghorn's APIs and controllers using the Operator SDK will
simplify the codebase and provide the right level of abstraction to
support rapid ongoing development of Foghorn.